requirements.txt
streamlit
pandas
numpy
requests
plotly
matplotlib
python-dotenv

